requirements.txt
base58==2.0.0
Django==3.0.5
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0